requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
asyncpg>=0.24.0
pydantic>=1.8.2
python-dotenv>=0.19.0
alembic>=1.7.1
psycopg2-binary>=2.9.1
pytest>=7.0.0
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5 